separate the dream structuring algorithm into its own file.
Add the base class for structuring algorithms.
