+++ angr/analyses/decompiler/__init__.py
@@ -1,5 +1,4 @@
 from .region_identifier import RegionIdentifier
-from .structuring import Structurer
 from .structured_codegen import CStructuredCodeGenerator, ImportSourceCode
 from .clinic import Clinic
 from .region_simplifiers import RegionSimplifier
@@ -8,6 +7,7 @@ from .decompilation_options import options, options_by_category
 from .block_simplifier import BlockSimplifier
 from .callsite_maker import CallSiteMaker
 from .ail_simplifier import AILSimplifier
+from . import structuring
 from . import optimization_passes
 
 StructuredCodeGenerator = CStructuredCodeGenerator
