+++ angr/analyses/decompiler/structuring/__init__.py
@@ -1,2 +1,2 @@
-from .structurer import Structurer
+from .dream import DreamStructurer
 from .recursive_structurer import RecursiveStructurer
